+++ test/cypress/tests/utils/flipper.cy.ts
@@ -46,10 +46,10 @@ class SomePlugin {
 }
 
 describe('Flipper', () => {
-  it('should prevent plugins event handlers from being called while keyboard navigation', () => {
-    const ARROW_DOWN_KEY_CODE = 40;
-    const ENTER_KEY_CODE = 13;
+  const ARROW_DOWN_KEY_CODE = 40;
+  const ENTER_KEY_CODE = 13;
 
+  it('should prevent plugins event handlers from being called while keyboard navigation', () => {
     const sampleText = 'sample text';
 
     cy.createEditor({
@@ -100,5 +100,41 @@ describe('Flipper', () => {
       .trigger('keydown', { keyCode: ENTER_KEY_CODE });
 
     expect(SomePlugin.pluginInternalKeydownHandler).to.have.not.been.called;
+  });
+
+  it('should not flip when shift key is pressed', () => {
+    cy.createEditor({
+      data: {
+        blocks: [
+          {
+            type: 'paragraph',
+            data: {
+              text: 'Workspace in classic editors is made of a single contenteditable element, used to create different HTML markups. Editor.js workspace consists of separate Blocks: paragraphs, headings, images, lists, quotes, etc. Each of them is an independent contenteditable element (or more complex structure) provided by Plugin and united by Editor\'s Core.',
+            },
+          },
+        ],
+      },
+      autofocus: true,
+    });
+
+    cy.get('[data-cy=editorjs]')
+      .get('.ce-paragraph')
+      .as('paragraph')
+      .selectTextByOffset([0, 10])
+      .wait(200);
+
+    cy.get('@paragraph')
+      .trigger('keydown', { keyCode: ARROW_DOWN_KEY_CODE,
+        shiftKey: true });
+
+    // eslint-disable-next-line cypress/require-data-selectors
+    cy.get('[data-cy="inline-toolbar"]')
+      .get('.ce-popover--opened')
+      .as('popover')
+      .should('exist');
+
+    cy.get('@popover')
+      .get('.ce-popover-item--focused')
+      .should('not.exist');
   });
 });